+++ src/game/game.cpp
@@ -149,7 +149,7 @@ void Game::start()
 
 	if ( GameState::tick == 0 && !GameState::initialSave )
 	{
-		Global::cfg->set( "DaysToNextAutoSave", 0.0 );
+		Global::cfg->set( "DaysToNextAutoSave", 0 );
 		autoSave();
 		Global::cfg->set( "Pause", true );
 		signalPause( true );
@@ -434,7 +434,7 @@ void Game::processPlants()
 
 void Game::autoSave()
 {
-	int daysToNext = Global::cfg->get_or_default<double>( "DaysToNextAutoSave" , 0 );
+	int daysToNext = Global::cfg->get_or_default<int>( "DaysToNextAutoSave" , 0 );
 
 	if ( daysToNext == 0 )
 	{
@@ -451,7 +451,7 @@ void Game::autoSave()
 			Global::cfg->set( "Pause", false );
 		}
 
-		Global::cfg->set( "DaysToNextAutoSave", Global::cfg->get<double>( "AutoSaveInterval" ) - 1 );
+		Global::cfg->set( "DaysToNextAutoSave", Global::cfg->get<int>( "AutoSaveInterval" ) - 1 );
 	}
 	else
 	{
